src/scenario-04-meai/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
openai>=1.0.0
vibevoice[streamingtts] @ git+https://github.com/microsoft/VibeVoice.git
torch>=2.0.0
soundfile>=0.12.0
nemo_toolkit[asr]>=2.0.0
faster-whisper>=1.0.0
numpy
